Rename the package from ramanujan to ramanujantools
