+++ ramanujantools/cmf/known_cmfs/known_cmfs.py
@@ -1,9 +1,9 @@
 import sympy as sp
 from sympy.abc import a, b, c, x, y
 
-from ramanujan import Matrix
-from ramanujan.cmf import CMF
-from ramanujan.cmf.ffbar import FFbar
+from ramanujantools import Matrix
+from ramanujantools.cmf import CMF
+from ramanujantools.cmf.ffbar import FFbar
 
 c0, c1, c2, c3 = sp.symbols("c:4")
