+++ ramanujantools/delta.py
