+++ ramanujantools/generic_polynomial.py
