+++ ramanujantools/pcf/pcf_from_matrix_test.py
@@ -1,11 +1,11 @@
 from sympy.abc import c, n, x, y
 
-from ramanujan import Matrix
-from ramanujan.pcf import PCF, PCFFromMatrix
+from ramanujantools import Matrix
+from ramanujantools.pcf import PCF, PCFFromMatrix
 
 
 def test_pcf_from_matrix_parametric():
-    from ramanujan.cmf.known_cmfs import cmf1, c0, c1, c2, c3
+    from ramanujantools.cmf.known_cmfs import cmf1, c0, c1, c2, c3
 
     cmf = cmf1().subs([[c0, 0], [c1, 1], [c2, 1], [c3, c]])
     matrix = cmf.trajectory_matrix({x: 1, y: 1}, {x: 1, y: 1})
@@ -13,7 +13,7 @@ def test_pcf_from_matrix_parametric():
 
 
 def test_pcf_from_matrix_relative_limit():
-    from ramanujan.cmf.known_cmfs import cmf1, c0, c1, c2, c3
+    from ramanujantools.cmf.known_cmfs import cmf1, c0, c1, c2, c3
 
     cmf = cmf1().subs([[c0, 0], [c1, 1], [c2, 1], [c3, c]])
     matrix = cmf.trajectory_matrix({x: 1, y: 1}, {x: 1, y: 1})
@@ -22,7 +22,7 @@ def test_pcf_from_matrix_relative_limit():
 
 
 def test_as_pcf_equivalent():
-    from ramanujan.cmf.known_cmfs import cmf1, c0, c1, c2, c3
+    from ramanujantools.cmf.known_cmfs import cmf1, c0, c1, c2, c3
 
     cmf = cmf1().subs([[c0, 0], [c1, 1], [c2, 1], [c3, c]])
     matrix = cmf.trajectory_matrix({x: 1, y: 1}, {x: 1, y: 1})
