+++ requirements.txt
@@ -1,4 +1,4 @@
 mpmath>=1.3.0
 multimethod>=1.10
-scipy>=1.10.1
+pytest>=8.2.0
 sympy>=1.11.1
